+++ apps/client/src/plugins/LayerSwitcher/components/LayerItem.jsx
@@ -407,6 +407,7 @@ function LayerItem({
             />
             <ListItemText
               primary={layerCaption}
+              sx={{ alignSelf: "center" }}
               slotProps={{
                 primary: {
                   variant: "body1",