+++ apps/client/src/plugins/LayerSwitcher/components/LayerItemInfo.jsx
@@ -85,9 +85,9 @@ export default function LayerItemInfo({ layer, app, chapters }) {
   };
 
   return (
-    <>
+    <Box id="layer-item-details-info">
       {hasInfo() ? (
-        <Box id="layer-item-details-info">
+        <>
           {/* Infotext */}
           {layerInfo.infoText && (
             <>
@@ -120,12 +120,10 @@ export default function LayerItemInfo({ layer, app, chapters }) {
             />
           )}
           {renderChapterLinks()}
-        </Box>
+        </>
       ) : (
-        <Typography id="layer-item-details-info">
-          Ingen information tillgänglig
-        </Typography>
+        <Typography variant="body2">Ingen information tillgänglig</Typography>
       )}
-    </>
+    </Box>
   );
 }